+++ src/crawler.ts
@@ -11,3 +11,7 @@ export function searchUmin(uminId: string) {
   const response = UrlFetchApp.fetch('https://upload.umin.ac.jp/cgi-open-bin/ctr/index.cgi', options).getContentText('UTF-8');
   return response
 }
+
+export function readRecpt(recptNo: string) {
+  return UrlFetchApp.fetch('https://upload.umin.ac.jp/cgi-open-bin/ctr/ctr_view.cgi?recptno=' + recptNo).getContentText('UTF-8');
+}
