+++ src/umin.ts
@@ -15,3 +15,26 @@ export function getRecptNoFromData(data: string): string | undefined {
   }
   return recptNo;
 }
+
+interface RecptDataType {
+  target?: string
+  intervention? :string
+}
+
+export function getRecptData(response: string): RecptDataType {
+  var root = getHtmlRootElement(response);
+  var data: RecptDataType = {};
+  if (root) {
+    var tds = getHtmlElementsByTagName(root, 'td');
+    for (var i = 0; i < tds.length; i++) {
+      if (tds[i].text.indexOf('対象疾患名/Condition') != -1) {
+        data.target = tds[i + 1].text;
+      }
+      if (tds[i].text.indexOf('介入1/Interventions/Control_1') != -1) {
+        data.intervention = tds[i + 1].text;
+        break;
+      }
+    }
+  }
+  return data;
+}
