+++ tests/umin.test.ts
@@ -1,7 +1,7 @@
 import { assert } from "chai"
-//import axios from 'axios'
+import axios from 'axios'
 import fs from 'fs'
-import { getRecptNoFromData } from "../src/umin"
+import { getRecptNoFromData, getRecptData } from "../src/umin"
 
 /*
 async function searchUmin(uminId: string): Promise<string> {
@@ -27,6 +27,19 @@ describe('getRecptNo', () => {
       const data = fs.readFileSync(`tests/fixtures/umin_${uminId}.html`).toString()
       const recptNo = getRecptNoFromData(data)
       assert.equal(recptNo, 'R000031865')
+    })
+  })
+})
+
+describe('getUminData', () => {
+  context('valid recptNo', () => {
+    const recptNo = 'R000031865'
+    it('returns valid data', () => {
+      // curl 'https://upload.umin.ac.jp/cgi-open-bin/ctr/ctr_view.cgi?recptno=R000031865' > tests/fixtures/umin_recpt_R000031865.html
+      const html = fs.readFileSync(`tests/fixtures/umin_recpt_${recptNo}.html`).toString()
+      const recptData = getRecptData(html)
+      assert.equal(recptData.target, '急性白血病')
+      assert.equal(recptData.intervention, 'FLAMEL regimen:リン酸フルダラビン(30 mg/square meter×1回/日を4日間)と、シタラビン（2 g/square meter×1回/日を2日間）、メルファラン（60 mg/square を3日間）、全身照射（3 Gy 1ｘ1/日）\r移植片は骨髄、末梢血幹細胞、臍帯血のいずれも選択できる。原則的に，タクロリムスまたはシクロスポリンとメソトレキサートを用いたGVHD予防を行う。')
     })
   })
 })
